Add files via upload
This what would happen if, a kid gets a new dog. I'm trying to teach kids to take care of there dogs. As well as trying to get people to be more aware of there dogs. That dogs are a big responsibility not just a cute pet. This will have less people losing dogs and taking care of there dogs instead.